Revert to two file system
